Consolidate closure calls with command calls
Refactor command call handling to avoid code duplication with closure
calls.
